[Refact] #32 - 지도 컴포넌트 ref 전환 및 코드 리팩토링
