+++ apps/service/src/pages/main/_components/map/MainMap.tsx
@@ -1,22 +1,7 @@
-import { useState, useEffect } from "react";
 import * as S from "./MainMap.styled";
+import { useState, useEffect, useRef } from "react";
 import { useGetBoothsLocation } from "@hooks/apis/booth";
-
-let mapInstance: naver.maps.Map | undefined;
-
-const loadScript = (src: string, callback: () => void) => {
-  if (document.querySelector(`script[src="${src}"]`)) {
-    callback();
-    return;
-  }
-  const script = document.createElement("script");
-  script.type = "text/javascript";
-  script.src = src;
-  script.onload = () => {
-    callback();
-  };
-  document.head.appendChild(script);
-};
+import { loadScript } from "@utils/loadScript";
 
 const MainMap = ({
   latitude,
@@ -25,45 +10,49 @@ const MainMap = ({
   latitude: number;
   longitude: number;
 }) => {
-  const [isScriptLoaded, setScriptLoaded] = useState(false);
   const { data: locations } = useGetBoothsLocation();
+
+  const [isScriptLoaded, setScriptLoaded] = useState(false);
+  const mapRef = useRef<naver.maps.Map | null>(null);
+
   console.log(locations);
 
   useEffect(() => {
-    if (typeof window.naver !== "undefined") {
-      setScriptLoaded(true);
-    } else {
+    if (typeof window.naver === "undefined") {
       loadScript(
         `https://oapi.map.naver.com/openapi/v3/maps.js?ncpKeyId=${
           import.meta.env.VITE_NAVER_MAP_KEY
-        }`,
-        () => setScriptLoaded(true)
+        }`
       );
     }
+    setScriptLoaded(true);
   }, []);
 
   useEffect(() => {
-    if (isScriptLoaded) {
-      setTimeout(() => {
-        mapInstance = new naver.maps.Map("map", {
-          center: new naver.maps.LatLng(latitude, longitude),
-          zoom: 16,
-          zoomControl: true,
-          zoomControlOptions: {
-            style: naver.maps.ZoomControlStyle.SMALL,
-            position: naver.maps.Position.TOP_RIGHT,
-          },
-        });
+    if (!isScriptLoaded) return;
 
-        // new naver.maps.Marker({
-        //   position: new naver.maps.LatLng(latitude, longitude),
-        //   map: mapInstance,
-        // });
-      }, 500);
-    }
+    setTimeout(() => {
+      mapRef.current = new naver.maps.Map("map", {
+        center: new naver.maps.LatLng(latitude, longitude),
+        zoom: 16,
+        zoomControl: true,
+        zoomControlOptions: {
+          style: naver.maps.ZoomControlStyle.SMALL,
+          position: naver.maps.Position.TOP_RIGHT,
+        },
+      });
+    }, 500);
   }, [isScriptLoaded, latitude, longitude]);
 
   return <div id="map" css={S.getWrapper}></div>;
 };
 
 export default MainMap;
+
+/**
+ *  TODO
+ *      new naver.maps.Marker({
+        position: new naver.maps.LatLng(latitude, longitude),
+        map: mapRef.current,
+      });
+ */
